requirements.txt
playwright
requests
python-dotenv
google-auth
google-api-python-client